+++ apps/dev/src/registries.tsx
@@ -134,6 +134,7 @@ export const gameRegistry: GameRegistry = {
   "extraction-map": () => import("./demo/mapDemo").then((module) => module.mapDemoGame),
   "notification-center": () => import("./demo/notificationDemo").then((module) => module.notificationDemoGame),
   "coach-marks": () => import("./demo/coachMarksDemo").then((module) => module.coachMarksDemoGame),
+  "damage-direction": () => import("./demo/damageDirectionDemo").then((module) => module.damageDirectionDemoGame),
   "day-night": () => import("./demo/dayNightDemo").then((module) => module.dayNightDemoGame),
   "status-effects": () => import("./demo/statusEffectDemo").then((module) => module.statusEffectDemoGame),
   "cutscene": () => import("./demo/cutsceneDemo").then((module) => module.cutsceneDemoGame),